Web crud endpoints

+++ persistence/src/main/java/com/redberrystudios/whatsfordinner/board/BoardMongoRepository.java
@@ -1,24 +1,25 @@
 package com.redberrystudios.whatsfordinner.board;
 
-import static com.mongodb.client.model.Filters.eq;
-import static com.mongodb.client.model.Filters.in;
-
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.UpdateOptions;
 import com.redberrystudios.whatsfordinner.MongoRepository;
-import com.redberrystudios.whatsfordinner.day.DayMongoRepository;
+import com.redberrystudios.whatsfordinner.group.DayElementEntity;
 import com.redberrystudios.whatsfordinner.group.GroupEntity;
 import com.redberrystudios.whatsfordinner.group.GroupMongoRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
+
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Repository;
+
+import static com.mongodb.client.model.Filters.eq;
+import static com.mongodb.client.model.Filters.in;
 
 @Repository
-public class BoardMongoRepository extends MongoRepository {
+public class BoardMongoRepository extends MongoRepository<BoardEntity, Long> {
 
   private static final String COLLECTION_NAME = "boards";
 
@@ -26,37 +27,48 @@ public class BoardMongoRepository extends MongoRepository {
 
   private GroupMongoRepository groupMongoRepository;
 
-  private DayMongoRepository dayMongoRepository;
-
   @Autowired
-  public BoardMongoRepository(MongoDatabase db, GroupMongoRepository groupMongoRepository, DayMongoRepository dayMongoRepository) {
+  public BoardMongoRepository(MongoDatabase db, GroupMongoRepository groupMongoRepository) {
     super(db);
     collection = database.getCollection(COLLECTION_NAME, BoardEntity.class);
-    
+
     this.groupMongoRepository = groupMongoRepository;
-    this.dayMongoRepository = dayMongoRepository;
   }
 
-  public void save(BoardEntity board) {
-    collection.replaceOne(eq("_id", board.getId()),
-        board,
-        new UpdateOptions().upsert(true));
+  public Long save(BoardEntity board) {
+    if (board == null) {
+      throw new IllegalArgumentException("BoardEntity to save is null!");
+    }
+
+    collection.replaceOne(eq("_id", board.getId()), board, new UpdateOptions().upsert(true));
+
+    return board.getId();
   }
 
-  public void delete(BoardEntity board) {
+  public Long delete(BoardEntity board) {
+    if (board == null) {
+      throw new IllegalArgumentException("BoardEntity to delete is null!");
+    }
+
     collection.deleteOne(eq("_id", board.getId()));
+    return board.getId();
   }
 
   public BoardEntity find(Long boardId) {
+    if (boardId == null) {
+      return null;
+    }
     return collection.find(eq("_id", boardId)).first();
   }
 
   public BoardEntity findByGroup(Long groupId, Long boardId) {
+    if (groupId == null || boardId == null) {
+      return null;
+    }
     GroupEntity groupEntity = groupMongoRepository.find(groupId);
 
-    Boolean isValidBoardId = groupEntity.getDays()
-        .stream()
-        .anyMatch(dayId -> dayMongoRepository.find(dayId).getBoards().contains(boardId));
+    Boolean isValidBoardId = groupEntity.getDays().stream()
+        .anyMatch(day -> day.getBoards().contains(boardId));
 
     if (isValidBoardId) {
       return collection.find(eq("_id", boardId)).first();
@@ -66,11 +78,14 @@ public class BoardMongoRepository extends MongoRepository {
   }
 
   public List<BoardEntity> findAllByGroup(Long groupId) {
+    if (groupId == null) {
+      return new ArrayList<>();
+    }
+
     GroupEntity groupEntity = groupMongoRepository.find(groupId);
 
-    List<Long> boardIds = groupEntity.getDays()
-        .stream()
-        .map(dayId -> dayMongoRepository.find(dayId).getBoards())
+    List<Long> boardIds = groupEntity.getDays().stream()
+        .map(DayElementEntity::getBoards)
         .flatMap(Collection::stream)
         .collect(Collectors.toList());
