+++ persistence/src/main/java/com/redberrystudios/whatsfordinner/group/GroupMongoRepository.java
@@ -1,7 +1,5 @@
 package com.redberrystudios.whatsfordinner.group;
 
-import static com.mongodb.client.model.Filters.eq;
-
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.UpdateOptions;
@@ -9,8 +7,13 @@ import com.redberrystudios.whatsfordinner.MongoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
+import java.util.ArrayList;
+import java.util.List;
+
+import static com.mongodb.client.model.Filters.eq;
+
 @Repository
-public class GroupMongoRepository extends MongoRepository {
+public class GroupMongoRepository extends MongoRepository<GroupEntity, Long> {
 
   private static final String COLLECTION_NAME = "groups";
 
@@ -22,21 +25,51 @@ public class GroupMongoRepository extends MongoRepository {
     collection = database.getCollection(COLLECTION_NAME, GroupEntity.class);
   }
 
-  public void save(GroupEntity group) {
-    collection.replaceOne(eq("_id", group.getId()),
-        group,
-        new UpdateOptions().upsert(true));
+  public Long save(GroupEntity group) {
+    if (group == null) {
+      throw new IllegalArgumentException("GroupEntity to save is null!");
+    }
+
+    collection.replaceOne(eq("_id", group.getId()), group, new UpdateOptions().upsert(true));
+
+    return group.getId();
   }
 
-  public void delete(GroupEntity group) {
+  public Long delete(GroupEntity group) {
+    if (group == null) {
+      throw new IllegalArgumentException("GroupEntity to delete is null!");
+    }
+
     collection.deleteOne(eq("_id", group.getId()));
+    return group.getId();
   }
 
   public GroupEntity find(Long groupId) {
+    if (groupId == null) {
+      return null;
+    }
+
     return collection.find(eq("_id", groupId)).first();
   }
 
   public GroupEntity findByJoinToken(String joinToken) {
+    if (joinToken == null) {
+      return null;
+    }
+
     return collection.find(eq("joinToken", joinToken)).first();
   }
+
+  public GroupEntity findByMember(Long memberId) {
+    if (memberId == null) {
+      return null;
+    }
+
+    return collection.find(eq("members", memberId)).first();
+  }
+
+  public List<GroupEntity> findAll() {
+    return collection.find().into(new ArrayList<>());
+  }
+
 }
