+++ persistence/src/main/java/com/redberrystudios/whatsfordinner/member/MemberMongoRepository.java
@@ -1,22 +1,21 @@
 package com.redberrystudios.whatsfordinner.member;
 
+import static com.mongodb.client.model.Filters.eq;
+import static com.mongodb.client.model.Filters.in;
+
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.UpdateOptions;
 import com.redberrystudios.whatsfordinner.MongoRepository;
 import com.redberrystudios.whatsfordinner.group.GroupEntity;
 import com.redberrystudios.whatsfordinner.group.GroupMongoRepository;
+import java.util.ArrayList;
+import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import static com.mongodb.client.model.Filters.eq;
-import static com.mongodb.client.model.Filters.in;
-
 @Repository
-public class MemberMongoRepository extends MongoRepository {
+public class MemberMongoRepository extends MongoRepository<MemberEntity, Long> {
 
   private static final String COLLECTION_NAME = "members";
 
@@ -32,21 +31,38 @@ public class MemberMongoRepository extends MongoRepository {
     this.groupMongoRepository = groupMongoRepository;
   }
 
-  public void save(MemberEntity member) {
-    collection.replaceOne(eq("_id", member.getId()),
-        member,
-        new UpdateOptions().upsert(true));
+  public Long save(MemberEntity member) {
+    if (member == null) {
+      throw new IllegalArgumentException("MemberEntity to save is null!");
+    }
+
+    collection.replaceOne(eq("_id", member.getId()), member, new UpdateOptions().upsert(true));
+
+    return member.getId();
   }
 
-  public void delete(MemberEntity member) {
+  public Long delete(MemberEntity member) {
+    if (member == null) {
+      throw new IllegalArgumentException("MemberEntity to delete is null!");
+    }
+
     collection.deleteOne(eq("_id", member.getId()));
+    return member.getId();
   }
 
   public MemberEntity find(Long memberId) {
+    if (memberId == null) {
+      return null;
+    }
+
     return collection.find(eq("_id", memberId)).first();
   }
 
   public MemberEntity findByGroup(Long groupId, Long memberId) {
+    if (groupId == null || memberId == null) {
+      return null;
+    }
+
     GroupEntity groupEntity = groupMongoRepository.find(groupId);
 
     if (groupEntity.getMembers().contains(memberId)) {
@@ -57,7 +73,19 @@ public class MemberMongoRepository extends MongoRepository {
   }
 
   public List<MemberEntity> findAllByGroup(Long groupId) {
+    if (groupId == null) {
+      return new ArrayList<>();
+    }
+
     GroupEntity groupEntity = groupMongoRepository.find(groupId);
     return collection.find(in("_id", groupEntity.getMembers())).into(new ArrayList<>());
+  }
+
+  public MemberEntity findByEmail(String email) {
+    if (email == null) {
+      return null;
+    }
+
+    return collection.find(eq("email", email)).first();
   }
 }
